BFD Retrieval

+++ src/main/java/com/revenueforecast/repository/BFDRepository.java
@@ -3,5 +3,5 @@ package com.revenueforecast.repository;
 import com.revenueforecast.model.BFD;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BFDRepository extends JpaRepository<BFD, Long> {
+public interface BFDRepository extends JpaRepository<BFD, Integer> {
 }
